archivelist/requirements.txt
blinker==1.9.0
blurhash==1.1.4
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
decorator==5.2.1
Flask==3.1.0
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.2
Mastodon.py==2.0.1
python-dateutil==2.9.0.post0
python-magic==0.4.27
requests==2.32.3
six==1.17.0
urllib3==2.4.0
watchdog==6.0.0
Werkzeug==3.1.3
